added tests and support for big values of sort. Also made sure we do …

+++ src/main/java/no/sikt/sws/models/opensearch/DocDto.java
@@ -47,7 +47,7 @@ public class DocDto extends Dto {
     public JsonNode source;
 
     @JsonProperty("sort")
-    public List<Integer> sort;
+    public List<Long> sort;
 
 
     public DocDto() {
